requirements.txt
ftfy==5.5.1
gensim==3.7.1
numpy==1.17
pandas==0.24.2
pytorch-transformers==1.1.0
scipy==1.2.1
scispacy==0.2.3
scikit-learn==0.20.3
spacy==2.1.8
tensorflow-gpu==1.14
tqdm
cchardet
overrides==3.1.0
-e allennlp/
p_tqdm
lime
captum
setuptools